+++ src/gitlab.py
@@ -1,10 +1,10 @@
 # encoding: utf-8
 import sys
 import argparse
-from workflow import Workflow, ICON_WEB, ICON_WARNING, ICON_INFO, web, PasswordNotFound
+from workflow import Workflow3, ICON_WEB, ICON_WARNING, ICON_INFO, web, PasswordNotFound
 from workflow.background import run_in_background, is_running
 
-__version__ = '1.1.0'
+__version__ = '1.2.0'
 
 log = None
 
@@ -77,14 +77,22 @@ def main(wf):
 
     projects = wf.cached_data('projects', None, max_age=0)
 
+    if wf.update_available:
+        # Add a notification to top of Script Filter results
+        wf.add_item('New version available',
+                    'Action this item to install the update',
+                    autocomplete='workflow:update',
+                    icon=ICON_INFO)
+
     # Start update script if cached data is too old (or doesn't exist)
-    if not wf.cached_data_fresh('projects', max_age=600):
+    if not wf.cached_data_fresh('projects', max_age=3600):
         cmd = ['/usr/bin/python', wf.workflowfile('update.py')]
         run_in_background('update', cmd)
 
     # Notify the user if the cache is being updated
     if is_running('update'):
-        wf.add_item('Updating project list via GitLab',
+        wf.rerun = 1
+        wf.add_item('Updating project list via GitLab...',
                     valid=False,
                     icon=ICON_INFO)
 
@@ -104,13 +112,17 @@ def main(wf):
                     subtitle=project['path_with_namespace'],
                     arg=project['web_url'],
                     valid=True,
-                    icon=ICON_WEB)
+                    icon=None)
 
     # Send the results to Alfred as XML
     wf.send_feedback()
 
 
 if __name__ == u"__main__":
-    wf = Workflow()
+    wf = Workflow3(update_settings={
+        'github_slug': 'lukewaite/alfred-gitlab',
+        'version': __version__,
+        'frequency': 1
+    })
     log = wf.logger
     sys.exit(wf.run(main))