+++ CHANGELOG.md
@@ -5,7 +5,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-git stat
+## [v1.2.0] (2017-10-20)
+
+### Feature
+* Notify if alfred-workflow detects a new version of alfred-gitlab
+
+### Fixed
+* Refresh after background updating of GitLab projects list
+
+## [v1.1.0] (2017-10-20)
+* Enable background updating of GitLab projects list
 
 ## [v1.0.2] (2017-10-20)
 * Bump alfred-workflow to 1.28.1 to fix issues
@@ -17,8 +26,9 @@ git stat
 
 * Initial implementaiton of alfred-gitlab workflow
 
-[Unreleased]: https://github.com/lukewaite/alfred-gitlab/compare/v1.1.0...HEAD
-[v1.1.2]: https://github.com/lukewaite/alfred-gitlab/compare/v1.0.2...v1.1.0
+[Unreleased]: https://github.com/lukewaite/alfred-gitlab/compare/v1.2.0...HEAD
+[v1.2.0]: https://github.com/lukewaite/alfred-gitlab/compare/v1.1.0...v1.2.0
+[v1.1.0]: https://github.com/lukewaite/alfred-gitlab/compare/v1.0.2...v1.1.0
 [v1.0.2]: https://github.com/lukewaite/alfred-gitlab/compare/v1.0.1...v1.0.2
 [v1.0.1]: https://github.com/lukewaite/alfred-gitlab/compare/v1.0.0...v1.0.1
 [v1.0.0]: https://github.com/lukewaite/alfred-gitlab/compare/90b63639ac1d06f9a52c37afd3f9c1da37d6ebd2...v1.0.0